Move Lidgren-specific code into isolated namespace
