+++ Intersect.Network/Lidgren/LidgrenConnection.cs
@@ -7,7 +7,7 @@ using Intersect.Network.Packets;
 
 using Lidgren.Network;
 
-namespace Intersect.Network
+namespace Intersect.Network.Lidgren
 {
 
     public sealed class LidgrenConnection : AbstractConnection
