+++ Intersect.Server/Networking/Lidgren/ServerNetwork.cs
@@ -6,6 +6,7 @@ using System.Security.Cryptography;
 using Intersect.Logging;
 using Intersect.Network;
 using Intersect.Network.Events;
+using Intersect.Network.Lidgren;
 using Intersect.Server.Entities;
 
 using JetBrains.Annotations;
